handle frontend error at dashboard export
